+++ include/UECS/CmptTraits.hpp
@@ -9,6 +9,8 @@
 #include <string>
 
 namespace Ubpa::UECS {
+	class EntityMngr;
+
 	// run-time dynamic component traits
 	// size (> 0) is neccessary
 	// optional
@@ -23,20 +25,14 @@ namespace Ubpa::UECS {
 	public:
 		static constexpr std::size_t default_alignment = alignof(std::max_align_t);
 
-		CmptTraits();
-		CmptTraits(const CmptTraits& other);
-		CmptTraits(CmptTraits&& other) noexcept = default;
-		CmptTraits& operator=(const CmptTraits & other);
-		CmptTraits& operator=(CmptTraits&& other) noexcept = default;
-
 		CmptTraits& Clear();
 
 		CmptTraits& RegisterName(Type);
 		CmptTraits& RegisterTrivial(TypeID);
 		CmptTraits& RegisterSize(TypeID, std::size_t size);
 		CmptTraits& RegisterAlignment(TypeID, std::size_t alignment);
-		CmptTraits& RegisterDefaultConstructor(TypeID, std::function<void(void*)>);
-		CmptTraits& RegisterCopyConstructor(TypeID, std::function<void(void*,void*)>);
+		CmptTraits& RegisterDefaultConstructor(TypeID, std::function<void(void*, std::pmr::memory_resource*)>);
+		CmptTraits& RegisterCopyConstructor(TypeID, std::function<void(void*,const void*, std::pmr::memory_resource*)>);
 		CmptTraits& RegisterMoveConstructor(TypeID, std::function<void(void*,void*)>);
 		CmptTraits& RegisterMoveAssignment(TypeID, std::function<void(void*,void*)>);
 		CmptTraits& RegisterDestructor(TypeID, std::function<void(void*)>);
@@ -54,7 +50,7 @@ namespace Ubpa::UECS {
 		std::size_t Sizeof(TypeID) const;
 		std::size_t Alignof(TypeID) const;
 		void DefaultConstruct(TypeID, void* cmpt) const;
-		void CopyConstruct(TypeID, void* dst, void* src) const;
+		void CopyConstruct(TypeID, void* dst, const void* src) const;
 		void MoveConstruct(TypeID, void* dst, void* src) const;
 		void MoveAssign(TypeID, void* dst, void* src) const;
 		void Destruct(TypeID, void* cmpt) const;
@@ -72,6 +68,11 @@ namespace Ubpa::UECS {
 		void Deregister();
 
 	private:
+		friend class EntityMngr;
+		CmptTraits(std::pmr::memory_resource* world_rsrc);
+		CmptTraits(const CmptTraits& other, std::pmr::memory_resource* world_rsrc);
+		CmptTraits(CmptTraits&& other) noexcept;
+
 		// register all for Cmpt
 		// static_assert
 		// - is_default_constructible_v
@@ -90,11 +91,13 @@ namespace Ubpa::UECS {
 		std::unordered_map<TypeID, std::string_view> names;
 		std::unordered_map<TypeID, std::size_t> sizeofs;
 		std::unordered_map<TypeID, std::size_t> alignments;
-		std::unordered_map<TypeID, std::function<void(void*)>> default_constructors; // dst <- src
-		std::unordered_map<TypeID, std::function<void(void*,void*)>> copy_constructors; // dst <- src
-		std::unordered_map<TypeID, std::function<void(void*,void*)>> move_constructors; // dst <- src
-		std::unordered_map<TypeID, std::function<void(void*,void*)>> move_assignments; // dst <- src
+		std::unordered_map<TypeID, std::function<void(void*, std::pmr::memory_resource*)>> default_constructors; // dst <- src
+		std::unordered_map<TypeID, std::function<void(void*, const void*, std::pmr::memory_resource*)>> copy_constructors; // dst <- src
+		std::unordered_map<TypeID, std::function<void(void*, void*)>> move_constructors; // dst <- src
+		std::unordered_map<TypeID, std::function<void(void*, void*)>> move_assignments; // dst <- src
 		std::unordered_map<TypeID, std::function<void(void*)>> destructors;
+
+		std::pmr::memory_resource* world_rsrc;
 	};
 }