+++ src/core/ArchetypeCmptTraits.cpp
@@ -10,14 +10,14 @@
 
 using namespace Ubpa::UECS;
 
-void ArchetypeCmptTraits::CmptTrait::DefaultConstruct(void* cmpt) const {
+void ArchetypeCmptTraits::CmptTrait::DefaultConstruct(void* cmpt, std::pmr::memory_resource* world_rsrc) const {
 	if(default_ctor)
-		default_ctor(cmpt);
+		default_ctor(cmpt, world_rsrc);
 }
 
-void ArchetypeCmptTraits::CmptTrait::CopyConstruct(void* dst, void* src) const {
+void ArchetypeCmptTraits::CmptTrait::CopyConstruct(void* dst, const void* src, std::pmr::memory_resource* world_rsrc) const {
 	if (copy_ctor)
-		copy_ctor(dst, src);
+		copy_ctor(dst, src, world_rsrc);
 	else
 		std::memcpy(dst, src, size);
 }
@@ -25,8 +25,6 @@ void ArchetypeCmptTraits::CmptTrait::CopyConstruct(void* dst, void* src) const {
 void ArchetypeCmptTraits::CmptTrait::MoveConstruct(void* dst, void* src) const {
 	if (move_ctor)
 		move_ctor(dst, src);
-	else if (copy_ctor)
-		copy_ctor(dst, src);
 	else
 		std::memcpy(dst, src, size);
 }